requirements.txt
flask==2.0.1
crewai==0.61.3
python-dotenv==0.19.0
flask-cors==3.0.10
gunicorn==20.1.0 